% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgc.R
\name{estimate}
\alias{estimate}
\title{Estimate poisson parameters phi = alpha/beta and beta from sequence of times (T), number of centers (N), and number of recruits (n), across a collection of clinical trials.}
\usage{
estimate(T_seq, N_seq, n_seq, beta0 = NULL, beta_bnds = NULL)
}
\arguments{
\item{T_seq}{Sequence of recruitment times across trials.}

\item{N_seq}{Sequence of number of centers across trials.}

\item{n_seq}{Sequence of number of recruits across trials.}

\item{beta0}{Initial estimate for beta for coordinate descent. If not specified, beta0 is randomly chosen uniformly between beta_bnds.}

\item{beta_bnds}{Bounds for initalization of beta0 and optimization bounds. If not specified, defaults to 0 to 1.}
}
\value{
A list of output:
\itemize{
\item{phi_hat:} MLE of phi=alpha/beta
\item{beta_hat:} MLE of beta
\item{ll:} value of negative of log-profile-likelihood at MLE.
}
}
\description{
Estimation procedure calculates maximum likelihood estimates via profile likelihood on beta.
}
