% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgc.R
\name{get_mle}
\alias{get_mle}
\title{Estimate poisson parameters phi = alpha/beta and beta from sequence of times (T_seq), number of centers (N_seq), and number of recruits (n_seq), across a collection of clinical trials.}
\usage{
get_mle(
  T_seq,
  N_seq,
  n_seq,
  phi0 = NULL,
  beta0 = NULL,
  phi_bnds = NULL,
  beta_bnds = NULL,
  maxiter = 1000,
  eps_stop = 1e-07,
  verbose = TRUE,
  opts = NULL,
  tritter = floor(maxiter/100)
)
}
\arguments{
\item{T_seq}{Sequence of recruitment times across trials.}

\item{N_seq}{Sequence of number of centers across trials.}

\item{n_seq}{Sequence of number of recruits across trials.}

\item{phi0}{Initial estimate for phi for coordinate descent. If not specified, phi0 is randomly chosen uniformly between phi_bnds.}

\item{beta0}{Initial estimate for beta for coordinate descent. If not specified, beta0 is randomly chosen uniformly between beta_bnds.}

\item{phi_bnds}{Bounds for initalization of phi0 and optimization bounds. If not specified, defaults to 0 to 1.}

\item{beta_bnds}{Bounds for initalization of beta0 and optimization bounds. If not specified, defaults to 0 to 1.}

\item{maxiter}{Max number of coordinate descent iterations.}

\item{eps_stop}{Stopping criterion for coordinate descent. Stops when absolute value of difference between either estimate of phi or beta falls below eps_stop.}

\item{verbose}{Boolean of whether or not to print out coordinate descent optimization results.}

\item{opts}{Options for NLOPTR optimzation algorithm. Defaults to list(algorithm = 'NLOPT_LN_COBYLA', xtol_rel = 1e-07, maxeval = 100).}

\item{tritter}{If verbose=TRUE, how often to print out optimization results. Default is every 100 iterations.}
}
\value{
A list of output:
\itemize{
\item{est:} Vector of MLE estimates for phi=alpha/beta and beta.
\item{ll:} value of negative of log-likelihood at MLEs.
}
}
\description{
Estimation procedure calculates maximum likelihood estimates via coordinate descent on phi and beta. By default each iteration of coordinate descent uses NLOPTR non-linear optimization package using the Constrained Optimization BY Linear Approximation (COBYLA) method.
}
