% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgDownloadPrediXcan.R
\name{pgcDownloadPrediXcan}
\alias{pgcDownloadPrediXcan}
\title{Download the PrediXcan database}
\usage{
pgcDownloadPrediXcan(destDir = "PrediXcan")
}
\arguments{
\item{destDir}{character of length 1, the download directory}
}
\description{
Download the PrediXcan database
}
