% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predXUtils.R
\name{predXUtils}
\alias{predXUtils}
\alias{pgcDownloadPrediXcan}
\alias{pgcCombinePredxFiles}
\alias{pgcGetPredxWeights}
\title{Download the PrediXcan database}
\usage{
pgcDownloadPrediXcan(predDir = "PrediXcan", overwrite = FALSE)

pgcCombinePredxFiles(predDir)

pgcGetPredxWeights(predDB)
}
\arguments{
\item{predDir}{character of length 1, the download directory}

\item{predDB}{character of length 1, file path to a predXcan .db file}
}
\description{
Download the PrediXcan database
}
\section{Functions}{
\itemize{
\item \code{pgcDownloadPrediXcan}: Download predXcan to the given directory

\item \code{pgcCombinePredxFiles}: Combine tissues weights into single data.table object

\item \code{pgcGetPredxWeights}: Reads weight table from the given predXcan .db file
}}

