% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predXUtils.R
\name{predXUtils}
\alias{predXUtils}
\alias{pgcDownloadPrediXcan}
\alias{pgcCombinePredxFiles}
\alias{pgcListTissues}
\alias{pgcGetPredxWeights}
\alias{pgcSubsetTissues}
\title{Download the PrediXcan database}
\usage{
pgcDownloadPrediXcan(downloadDest = "PrediXcan", rmTar = TRUE)

pgcCombinePredxFiles(predDir, SubsetByTissue = FALSE)

pgcListTissues(combWts)

pgcGetPredxWeights(predDB)

pgcSubsetTissues(tis)
}
\arguments{
\item{downloadDest}{character of length 1, the download directory}

\item{rmTar}{logical of length 1, Should orginal tar file be removed?}

\item{predDir}{character of length 1, directory where db files have been stored}

\item{combWts}{character of length 1, directory where db files have been stored}

\item{predDB}{character of length 1, file path to a predXcan .db file}

\item{tis}{character of length 1, Tissue you want to subset combined weights file by}
}
\description{
Download the PrediXcan database
}
\section{Functions}{
\itemize{
\item \code{pgcDownloadPrediXcan}: Download predXcan to the given directory

\item \code{pgcCombinePredxFiles}: Combine tissues weights into single data.table object

\item \code{pgcListTissues}: Returns list of tissues involved in all prediXcan .db

\item \code{pgcGetPredxWeights}: Reads weight table from the given predXcan .db file

\item \code{pgcSubsetTissues}: Subsets combined weight table for a specific tissue or tissues
}}

